Conflicts:
	trw.cpp

+++ imageClass.cpp
@@ -15,7 +15,7 @@ image::image(vector<vector<int>> &input_array) {
         imageRow tempRow = imageRow();
         for(int x=0; x<this->width; ++x)
         {
-            tempRow.push_back((float)255*input_array[y][x]/NBR_CLASSES);
+            tempRow.push_back((double)255*input_array[y][x]/NBR_CLASSES);
         }
         this->data.push_back(tempRow);
     }
