+++ node.h
@@ -10,14 +10,14 @@ using namespace std;
 class node {
 
 public:
-    array<float, NBR_CLASSES> unaries;
+    array<double, NBR_CLASSES> unaries;
     int id;
     int next_unary;
 
     node(){};
     node(int id);
-    void addUnary(float value);
-    float get_min_unary();
+    void addUnary(double value);
+    double get_min_unary();
 };
 
 #endif
